[news + projections + micromorts]: update from artdgn/pages
 - Update age specific IFRs to recent Nature figures.
 - Update testing bais calculation to be windowed so that it adjusts
 to changing testing policies.
 - Add new page of deadly risk in micromorts map and its thumbnail.
 - Minor textual and graphical updates to maps and graphs.
 - Add world summary of S/I/R ratios in world stacked graphs.
 - Update methodology texts.
 - Add Makefile convenience target for transforming script file into
 calculated notebook.
